# openapi/delta-backend.openapi.yaml
openapi: 3.0.0
paths:
...
  /api/business/companies:
    post:
      operationId: BusinessProduct_createCompany
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductCompanyDto"
      responses:
        "201":
          description: ""
      security:
        - businessSession: []
          csrfToken: []
      summary: Create a product company
      tags:
        - Business Products
  /api/business/companies/{companyId}:
    put:
      operationId: BusinessProduct_updateCompany
      parameters:
        - name: companyId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductCompanyDto"
      responses:
        "200":
          description: ""
      security:
        - businessSession: []
          csrfToken: []
      summary: Update a product company
      tags:
        - Business Products
    delete:
      operationId: BusinessProduct_deleteCompany
      parameters:
        - name: companyId
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      security:
        - businessSession: []
          csrfToken: []
      summary: Delete an unused product company
      tags:
        - Business Products
  /api/business/categories:
    post:
      operationId: BusinessProduct_createCategory
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductCategoryDto"
      responses:
        "201":
          description: ""
      security:
        - businessSession: []
          csrfToken: []
      summary: Create a product category
      tags:
        - Business Products
  /api/business/categories/{categoryId}:
    put:
      operationId: BusinessProduct_updateCategory
      parameters:
        - name: categoryId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductCategoryDto"
      responses:
        "200":
          description: ""
      security:
        - businessSession: []
          csrfToken: []
      summary: Update a product category
      tags:
        - Business Products
    delete:
      operationId: BusinessProduct_deleteCategory
      parameters:
        - name: categoryId
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      security:
        - businessSession: []
          csrfToken: []
      summary: Delete an unused product category
      tags:
        - Business Products
  /api/business/products:
    post:
      operationId: BusinessProduct_createProduct
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductDto"
      responses:
        "201":
          description: ""
      security:
        - businessSession: []
          csrfToken: []
      summary: Create a product and optionally its company or category
      tags:
        - Business Products
  /api/business/products/{productId}:
    put:
      operationId: BusinessProduct_updateProduct
      parameters:
        - name: productId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductDto"
      responses:
        "200":
          description: ""
      security:
        - businessSession: []
          csrfToken: []
      summary: Update a product
      tags:
        - Business Products
    delete:
      operationId: BusinessProduct_deleteProduct
      parameters:
        - name: productId
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      security:
        - businessSession: []
          csrfToken: []
      summary: Delete a product
      tags:
        - Business Products
  /api/business/products/out-of-stock/{productId}:
    delete:
      operationId: BusinessProduct_deleteOutOfStockProduct
      parameters:
        - name: productId
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      security:
        - businessSession: []
          csrfToken: []
      summary: Delete a product only when it is out of stock
      tags:
        - Business Products
...
info:
  title: Delta Backend API
  description: Public catalog and product APIs, catalog access flows,
    authenticated business administration, and internal administration
    endpoints.
  version: 0.0.1
  contact: {}
...
servers: []
components:
  securitySchemes:
    businessSession:
      type: apiKey
      in: cookie
      name: business_session
      description: Business session cookie. Production uses __Host-business_session.
    catalogAccess:
      type: apiKey
      in: cookie
      name: catalog_access
      description: Catalog access session cookie.
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: CSRF token returned by business authentication.
    internalAdmin:
      type: apiKey
      in: header
      name: X-Backend-Admin-Token
      description: Token for the optional internal administration API.
  schemas:
...
    CreateProductCompanyDto:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
          maxLength: 160
        slug:
          type: string
        logoPath:
          type: string
          maxLength: 1024
        active:
          type: boolean
        sortOrder:
          type: number
          minimum: 0
      required:
        - id
        - name
    UpdateProductCompanyDto:
      type: object
      properties:
        name:
          type: string
          maxLength: 160
        slug:
          type: string
        logoPath:
          type: string
          maxLength: 1024
        active:
          type: boolean
        sortOrder:
          type: number
          minimum: 0
    CreateProductCategoryDto:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
          maxLength: 160
        slug:
          type: string
        companyIds:
          maxItems: 1000
          type: array
          items:
            type: string
        active:
          type: boolean
        sortOrder:
          type: number
          minimum: 0
      required:
        - id
        - name
        - companyIds
    UpdateProductCategoryDto:
      type: object
      properties:
        name:
          type: string
          maxLength: 160
        slug:
          type: string
        companyIds:
          maxItems: 1000
          type: array
          items:
            type: string
        active:
          type: boolean
        sortOrder:
          type: number
          minimum: 0
    CreateProductDto:
      type: object
      properties:
        productId:
          type: string
        name:
          type: string
          maxLength: 240
        sku:
          type: string
          maxLength: 160
        modelNumber:
          type: string
          maxLength: 160
        isNewCompany:
          type: boolean
        companyName:
          type: string
          maxLength: 160
        companyId:
          type: string
        isNewCategory:
          type: boolean
        categoryName:
          type: string
          maxLength: 160
        categoryId:
          type: string
        subcategoryId:
          type: string
        subcategoryName:
          type: string
          maxLength: 160
        subcategorySlug:
          type: string
        price:
          type: number
          minimum: 0
        currency:
          type: string
        discountPercentage:
          type: number
          minimum: 0
          maximum: 100
        inStock:
          type: boolean
        stockQuantity:
          type: number
          minimum: 0
        active:
          type: boolean
        shortDescription:
          type: string
          maxLength: 500
        description:
          type: string
          maxLength: 20000
        specifications:
          type: object
          additionalProperties: true
        thumbnailPath:
          type: string
          maxLength: 1024
        mainImagePath:
          type: string
          maxLength: 1024
        additionalImagePaths:
          maxItems: 20
          type: array
          items:
            type: string
            maxLength: 1024
        brochurePath:
          type: string
          maxLength: 1024
      required:
        - name
        - price
        - inStock
    UpdateProductDto:
      type: object
      properties:
        name:
          type: string
          maxLength: 240
        sku:
          type: string
          maxLength: 160
        modelNumber:
          type: string
          maxLength: 160
        companyId:
          type: string
        categoryId:
          type: string
        subcategoryId:
          type: string
        subcategoryName:
          type: string
          maxLength: 160
        subcategorySlug:
          type: string
        price:
          type: number
          minimum: 0
        currency:
          type: string
        discountPercentage:
          type: number
          minimum: 0
          maximum: 100
        inStock:
          type: boolean
        stockQuantity:
          type: number
          minimum: 0
        active:
          type: boolean
        shortDescription:
          type: string
          maxLength: 500
        description:
          type: string
          maxLength: 20000
        specifications:
          type: object
          additionalProperties: true
        thumbnailPath:
          type: string
          maxLength: 1024
        mainImagePath:
          type: string
          maxLength: 1024
        additionalImagePaths:
          maxItems: 20
          type: array
          items:
            type: string
            maxLength: 1024
        brochurePath:
          type: string
          maxLength: 1024
...